fix run_coveralls to do it in minimal way

+++ run_coveralls.py
@@ -2,18 +2,13 @@
 
 import os
 
-from contextlib import suppress
-from distutils.sysconfig import get_python_lib
 from subprocess import call
 
 
 if __name__ == '__main__':
-    # chdir to the site-packages directory so the report lists relative paths
     dot_coverage_path = os.path.join(os.getcwd(), '.coverage')
-    os.chdir(get_python_lib())
-    with suppress(OSError):
+    if os.path.isfile(dot_coverage_path):
         os.remove('.coverage')
-    os.symlink(dot_coverage_path, '.coverage')
 
     # create a report from the coverage data
     if 'TRAVIS' in os.environ:
